Remove util module in renderer
