+++ maplibre/src/render/eventually.rs
@@ -1,38 +1,4 @@
-use std::cmp::Ordering;
 use std::mem;
-
-/// A wrapper type that enables ordering floats. This is a work around for the famous "rust float
-/// ordering" problem. By using it, you acknowledge that sorting NaN is undefined according to spec.
-/// This implementation treats NaN as the "smallest" float.
-#[derive(Debug, Copy, Clone, PartialOrd)]
-pub struct FloatOrd(pub f32);
-
-impl PartialEq for FloatOrd {
-    fn eq(&self, other: &Self) -> bool {
-        if self.0.is_nan() && other.0.is_nan() {
-            true
-        } else {
-            self.0 == other.0
-        }
-    }
-}
-
-impl Eq for FloatOrd {}
-
-#[allow(clippy::derive_ord_xor_partial_ord)]
-impl Ord for FloatOrd {
-    fn cmp(&self, other: &Self) -> Ordering {
-        self.0.partial_cmp(&other.0).unwrap_or_else(|| {
-            if self.0.is_nan() && !other.0.is_nan() {
-                Ordering::Less
-            } else if !self.0.is_nan() && other.0.is_nan() {
-                Ordering::Greater
-            } else {
-                Ordering::Equal
-            }
-        })
-    }
-}
 
 /// Wrapper around a resource which can be initialized or uninitialized.
 /// Uninitialized resourced can be initialized by calling [`Eventually::initialize()`].
@@ -87,6 +53,13 @@ impl<T> Eventually<T> {
 
     pub fn take(&mut self) -> Eventually<T> {
         mem::replace(self, Eventually::Uninitialized)
+    }
+
+    pub fn ok_mut(&mut self) -> Option<&mut T> {
+        match self {
+            Eventually::Initialized(e) => Some(e),
+            Eventually::Uninitialized => None,
+        }
     }
 }
