+++ maplibre/src/render/resource/texture.rs
@@ -1,8 +1,8 @@
 //! Utility for a texture view which can either be created by a [`TextureView`](wgpu::TextureView)
 //! or [`SurfaceTexture`](wgpu::SurfaceTexture)
 
+use crate::render::eventually::HasChanged;
 use crate::render::settings::Msaa;
-use crate::render::util::HasChanged;
 use std::ops::Deref;
 
 /// Describes a [`TextureView`].
