+++ maplibre/src/util/math.rs
@@ -345,6 +345,39 @@ where
     }
 }
 
+/// A wrapper type that enables ordering floats. This is a work around for the famous "rust float
+/// ordering" problem. By using it, you acknowledge that sorting NaN is undefined according to spec.
+/// This implementation treats NaN as the "smallest" float.
+#[derive(Debug, Copy, Clone, PartialOrd)]
+pub struct FloatOrd(pub f32);
+
+impl PartialEq for FloatOrd {
+    fn eq(&self, other: &Self) -> bool {
+        if self.0.is_nan() && other.0.is_nan() {
+            true
+        } else {
+            self.0 == other.0
+        }
+    }
+}
+
+impl Eq for FloatOrd {}
+
+#[allow(clippy::derive_ord_xor_partial_ord)]
+impl Ord for FloatOrd {
+    fn cmp(&self, other: &Self) -> Ordering {
+        self.0.partial_cmp(&other.0).unwrap_or_else(|| {
+            if self.0.is_nan() && !other.0.is_nan() {
+                Ordering::Less
+            } else if !self.0.is_nan() && other.0.is_nan() {
+                Ordering::Greater
+            } else {
+                Ordering::Equal
+            }
+        })
+    }
+}
+
 pub const fn div_away(lhs: i32, rhs: i32) -> i32 {
     if rhs < 0 {
         panic!("rhs must be positive")
